requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
coloredlogs==15.0.1
flatbuffers==25.12.19
humanfriendly==10.0
idna==3.11
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==1.23.5
onnx==1.12.0
onnx-simplifier==0.4.36
onnxruntime==1.13.1
packaging==26.0
pillow==11.3.0
protobuf==3.20.1
Pygments==2.19.2
pyreadline3==3.5.4
requests==2.32.5
rich==14.3.2
setuptools==80.9.0
sympy==1.14.0
torch==1.13.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
torchvision==0.14.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.45.1
opencv-python==4.8.0.74